# fix(schema): use editor for rendering schema export instead of LG code for performance COMPASS-8980