--- a/packages/compass-e2e-tests/tests/auto-connect.test.ts
+++ b/packages/compass-e2e-tests/tests/auto-connect.test.ts
@@ -308,9 +308,7 @@ describe('Automatically connecting from the command line', function () {
 
       await browser.execute(() => {
         // eslint-disable-next-line @typescript-eslint/no-var-requires
-        (require('electron').ipcRenderer as any).call(
-          'test:show-connect-window'
-        );
+        require('electron').ipcRenderer.call('test:show-connect-window');
       });
 
       // Switch to the other window