--- a/packages/compass-e2e-tests/tests/collection-schema-tab.test.ts
+++ b/packages/compass-e2e-tests/tests/collection-schema-tab.test.ts
@@ -111,8 +111,7 @@ describe('Collection schema tab', function () {
 
   describe('with the enableExportSchema feature flag enabled', function () {
     beforeEach(async function () {
-      // TODO(COMPASS-8819): remove web skip when defaulted true.
-      skipForWeb(this, "can't toggle features in compass-web");
+      // TODO(COMPASS-8819): remove when defaulted true
       await browser.setFeature('enableExportSchema', true);
     });
 
@@ -135,7 +134,14 @@ describe('Collection schema tab', function () {
 
       const exportSchemaContent = browser.$(Selectors.ExportSchemaOutput);
       await exportSchemaContent.waitForDisplayed();
-      const text = await browser.$(Selectors.ExportSchemaOutput).getText();
+
+      let text = '';
+      await browser.waitUntil(async function () {
+        text = await browser.getCodemirrorEditorText(
+          Selectors.ExportSchemaOutput
+        );
+        return text.includes('$schema:');
+      });
       const parsedText = JSON.parse(text);
       delete parsedText.$defs;
       expect(parsedText).to.deep.equal({