--- a/packages/compass-e2e-tests/tests/collection-schema-tab.test.ts
+++ b/packages/compass-e2e-tests/tests/collection-schema-tab.test.ts
@@ -135,7 +135,6 @@ describe('Collection schema tab', function () {
 
       const exportSchemaContent = browser.$(Selectors.ExportSchemaOutput);
       await exportSchemaContent.waitForDisplayed();
-
       const text = await browser.getCodemirrorEditorText(
         Selectors.ExportSchemaOutput
       );