--- a/packages/compass-schema/src/components/export-schema-modal.tsx
+++ b/packages/compass-schema/src/components/export-schema-modal.tsx
@@ -2,7 +2,7 @@ import React, { type ChangeEvent, useCallback } from 'react';
 import { connect } from 'react-redux';
 import {
   Button,
-  Code,
+  KeylineCard,
   ModalBody,
   ModalHeader,
   ModalFooter,
@@ -15,6 +15,7 @@ import {
   Label,
   CancelLoader,
 } from '@mongodb-js/compass-components';
+import { CodemirrorMultilineEditor } from '@mongodb-js/compass-editor';
 
 import type { RootState } from '../stores/store';
 import {
@@ -34,9 +35,15 @@ const contentContainerStyles = css({
   paddingBottom: spacing[400],
 });
 
-const codeStyles = css({
+const codeEditorContainerStyles = css({
   maxHeight: `${spacing[1600] * 4 - spacing[800]}px`,
   overflow: 'auto',
+});
+
+const codeStyles = css({
+  '& .cm-editor': {
+    paddingLeft: spacing[2],
+  },
 });
 
 const footerStyles = css({
@@ -137,15 +144,21 @@ const ExportSchemaModal: React.FunctionComponent<{
             />
           )}
           {exportStatus === 'complete' && (
-            <Code
-              id="export-schema-content"
-              data-testid="export-schema-content"
-              language="json"
-              className={codeStyles}
-              copyable={true}
-            >
-              {exportedSchema ?? 'Empty'}
-            </Code>
+            <KeylineCard className={codeEditorContainerStyles}>
+              <CodemirrorMultilineEditor
+                id="export-schema-content"
+                data-testid="export-schema-content"
+                language="json"
+                className={codeStyles}
+                copyable={true}
+                showAnnotationsGutter={false}
+                showLineNumbers={false}
+                formattable={false}
+                initialJSONFoldAll={false}
+                readOnly
+                text={exportedSchema ?? 'Empty'}
+              ></CodemirrorMultilineEditor>
+            </KeylineCard>
           )}
           {exportStatus === 'error' && errorMessage && (
             <ErrorSummary