COMPASS-8980
LeafyGreen's Code component gets laggy when given a large document (mostly from computing the dom highlighting). The exported schema json can get large, especially in the `Extended` format. This PR updates the export schema modal copyable output to be a code mirror read only editor. It renders it virtually and makes it a lot quicker to render.

Quick speed benchmarks on my machine, all collections are in Compass Data Sets, tested with the `Extended` format:
| namespace | render time before (ms) | render time after (ms) |
| - | - | - |
| NYC.buildings | 2616 | 65 |
| NYC.trees | 1979 | 50 |
| sample_airbnb.lightingsAndReviews | 15228 | 279 |

| before | after |
| - | - |
| <img width="627" alt="Screenshot 2025-02-18 at 2 10 40 PM" src="https://github.com/user-attachments/assets/01dbe268-996b-44d6-bc9d-32cf0fd01e93" /> | <img width="618" alt="Screenshot 2025-02-18 at 2 09 57 PM" src="https://github.com/user-attachments/assets/c9031f80-ec6b-4fa2-8415-ef79ad6c375b" /> |

^ There's a copy button on the after if the user hovers over it.